Improve modularity of AMMSwapPlatform
