+++ rotkehlchen/chain/ethereum/interfaces/ammswap/__init__.py
@@ -1,0 +1,6 @@
+from .ammswap import SUSHISWAP_TRADES_PREFIX, UNISWAP_TRADES_PREFIX
+
+__all__ = [
+    'UNISWAP_TRADES_PREFIX',
+    'SUSHISWAP_TRADES_PREFIX',
+]
