+++ rotkehlchen/chain/ethereum/interfaces/ammswap/typing.py
