+++ rotkehlchen/chain/ethereum/modules/sushiswap/__init__.py
@@ -1,5 +1,5 @@
-from .sushiswap import SUSHISWAP_EVENTS_PREFIX, SUSHISWAP_TRADES_PREFIX, Sushiswap  # noqa: F401
-from rotkehlchen.chain.ethereum.modules.ammswap.typing import (  # noqa: F401
+from .sushiswap import SUSHISWAP_EVENTS_PREFIX, Sushiswap  # noqa: F401
+from rotkehlchen.chain.ethereum.interfaces.ammswap.typing import (  # noqa: F401
     LiquidityPool as SushiswapPool,
     LiquidityPoolAsset as SushiswapPoolAsset,
     LiquidityPoolEvent as SushiswapPoolEvent,
@@ -8,7 +8,6 @@ from rotkehlchen.chain.ethereum.modules.ammswap.typing import (  # noqa: F401
 
 __all__ = [
     'SUSHISWAP_EVENTS_PREFIX',
-    'SUSHISWAP_TRADES_PREFIX',
     'SushiswapPool',
     'SushiswapPoolAsset',
     'SushiswapPoolEvent',
