+++ rotkehlchen/chain/ethereum/modules/sushiswap/graph.py
@@ -57,48 +57,3 @@ TOKEN_DAY_DATAS_QUERY = (
     }}}}
     """
 )
-
-# Get trades queries
-
-SWAPS_QUERY = (
-    """
-    swaps
-    (
-        first: $limit,
-        skip: $offset,
-        where: {{
-            to: $address,
-            timestamp_gte: $start_ts,
-            timestamp_lte: $end_ts,
-        }}
-    ) {{
-        transaction {{
-            swaps {{
-                id
-                logIndex
-                sender
-                to
-                timestamp
-                pair {{
-                    token0 {{
-                        id
-                        decimals
-                        name
-                        symbol
-                    }}
-                    token1 {{
-                        id
-                        decimals
-                        name
-                        symbol
-                    }}
-                }}
-                amount0In
-                amount0Out
-                amount1In
-                amount1Out
-            }}
-        }}
-    }}}}
-    """
-)
