+++ rotkehlchen/chain/ethereum/modules/uniswap/__init__.py
@@ -1,5 +1,5 @@
-from .uniswap import UNISWAP_EVENTS_PREFIX, UNISWAP_TRADES_PREFIX, Uniswap  # noqa: F401
-from rotkehlchen.chain.ethereum.modules.ammswap.typing import (  # noqa: F401
+from .uniswap import UNISWAP_EVENTS_PREFIX, Uniswap  # noqa: F401
+from rotkehlchen.chain.ethereum.interfaces.ammswap.typing import (  # noqa: F401
     LiquidityPool as UniswapPool,
     LiquidityPoolAsset as UniswapPoolAsset,
     LiquidityPoolEvent as UniswapPoolEvent,
@@ -8,7 +8,6 @@ from rotkehlchen.chain.ethereum.modules.ammswap.typing import (  # noqa: F401
 
 __all__ = [
     'UNISWAP_EVENTS_PREFIX',
-    'UNISWAP_TRADES_PREFIX',
     'UniswapPool',
     'UniswapPoolAsset',
     'UniswapPoolEvent',
