+++ rotkehlchen/chain/ethereum/modules/uniswap/utils.py
@@ -8,13 +8,13 @@ import requests
 from rotkehlchen.assets.asset import EthereumToken
 from rotkehlchen.chain.ethereum.contracts import EthereumContract
 from rotkehlchen.chain.ethereum.defi.zerionsdk import ZERION_ADAPTER_ADDRESS
-from rotkehlchen.chain.ethereum.modules.ammswap.utils import _decode_result
+from rotkehlchen.chain.ethereum.interfaces.ammswap.utils import _decode_result
 from rotkehlchen.chain.ethereum.typing import NodeName
 from rotkehlchen.constants.ethereum import ZERION_ABI
 from rotkehlchen.typing import ChecksumEthAddress
 from rotkehlchen.utils.misc import get_chunks
 
-from rotkehlchen.chain.ethereum.modules.ammswap.typing import LiquidityPool
+from rotkehlchen.chain.ethereum.interfaces.ammswap.typing import LiquidityPool
 
 if TYPE_CHECKING:
     from rotkehlchen.chain.ethereum.manager import EthereumManager
