+++ rotkehlchen/tests/unit/uniswap/test_get_balances_graph.py
@@ -1,6 +1,6 @@
 import pytest
 
-from rotkehlchen.chain.ethereum.modules.ammswap.typing import ProtocolBalance
+from rotkehlchen.chain.ethereum.interfaces.ammswap.typing import ProtocolBalance
 
 from .utils import (
     EXP_KNOWN_ASSETS_1,
